+++ spring-webmvc/src/main/java/org/springframework/web/servlet/handler/AbstractHandlerMapping.java
@@ -39,6 +39,7 @@ import org.springframework.util.Assert;
 import org.springframework.util.CollectionUtils;
 import org.springframework.util.PathMatcher;
 import org.springframework.web.HttpRequestHandler;
+import org.springframework.web.accept.ApiVersionStrategy;
 import org.springframework.web.context.request.WebRequestInterceptor;
 import org.springframework.web.context.request.async.WebAsyncManager;
 import org.springframework.web.context.request.async.WebAsyncUtils;
@@ -103,6 +104,8 @@ public abstract class AbstractHandlerMapping extends WebApplicationObjectSupport
 	private @Nullable CorsConfigurationSource corsConfigurationSource;
 
 	private CorsProcessor corsProcessor = new DefaultCorsProcessor();
+
+	private @Nullable ApiVersionStrategy versionStrategy;
 
 	private int order = Ordered.LOWEST_PRECEDENCE;  // default: same as non-Ordered
 
@@ -398,6 +401,23 @@ public abstract class AbstractHandlerMapping extends WebApplicationObjectSupport
 	}
 
 	/**
+	 * Configure a strategy to manage API versioning.
+	 * @param strategy the strategy to use
+	 * @since 7.0
+	 */
+	public void setApiVersionStrategy(@Nullable ApiVersionStrategy strategy) {
+		this.versionStrategy = strategy;
+	}
+
+	/**
+	 * Return the configured {@link ApiVersionStrategy} strategy.
+	 * @since 7.0
+	 */
+	public @Nullable ApiVersionStrategy getApiVersionStrategy() {
+		return this.versionStrategy;
+	}
+
+	/**
 	 * Specify the order value for this HandlerMapping bean.
 	 * <p>The default value is {@code Ordered.LOWEST_PRECEDENCE}, meaning non-ordered.
 	 * @see org.springframework.core.Ordered#getOrder()
@@ -519,6 +539,7 @@ public abstract class AbstractHandlerMapping extends WebApplicationObjectSupport
 	 */
 	@Override
 	public final @Nullable HandlerExecutionChain getHandler(HttpServletRequest request) throws Exception {
+		initApiVersion(request);
 		Object handler = getHandlerInternal(request);
 		if (handler == null) {
 			handler = getDefaultHandler();
@@ -561,6 +582,18 @@ public abstract class AbstractHandlerMapping extends WebApplicationObjectSupport
 		}
 
 		return executionChain;
+	}
+
+	private void initApiVersion(HttpServletRequest request) {
+		if (this.versionStrategy != null) {
+			Comparable<?> version = (Comparable<?>) request.getAttribute(API_VERSION_ATTRIBUTE);
+			if (version == null) {
+				version = this.versionStrategy.resolveParseAndValidateVersion(request);
+				if (version != null) {
+					request.setAttribute(API_VERSION_ATTRIBUTE, version);
+				}
+			}
+		}
 	}
 
 	/**
@@ -634,6 +667,7 @@ public abstract class AbstractHandlerMapping extends WebApplicationObjectSupport
 	 * @see #getAdaptedInterceptors()
 	 */
 	protected HandlerExecutionChain getHandlerExecutionChain(Object handler, HttpServletRequest request) {
+
 		HandlerExecutionChain chain = (handler instanceof HandlerExecutionChain handlerExecutionChain ?
 				handlerExecutionChain : new HandlerExecutionChain(handler));
 
@@ -647,6 +681,14 @@ public abstract class AbstractHandlerMapping extends WebApplicationObjectSupport
 				chain.addInterceptor(interceptor);
 			}
 		}
+
+		if (this.versionStrategy != null) {
+			Comparable<?> version = (Comparable<?>) request.getAttribute(API_VERSION_ATTRIBUTE);
+			if (version != null) {
+				chain.addInterceptor(new ApiVersionDeprecationHandlerInterceptor(this.versionStrategy, version));
+			}
+		}
+
 		return chain;
 	}
 
@@ -752,6 +794,17 @@ public abstract class AbstractHandlerMapping extends WebApplicationObjectSupport
 		@Override
 		public void handlePreFlight(HttpServletRequest request, HttpServletResponse response) throws IOException {
 			invokeCorsProcessor(request, response);
+		}
+	}
+
+
+	private record ApiVersionDeprecationHandlerInterceptor(
+			ApiVersionStrategy versionStrategy, Comparable<?> version) implements HandlerInterceptor {
+
+		@Override
+		public boolean preHandle(HttpServletRequest request, HttpServletResponse response, Object handler) {
+			this.versionStrategy.handleDeprecations(this.version, request, response);
+			return true;
 		}
 	}
